+++ aid.h
