+++ menu.h
@@ -1,137 +1,142 @@
-#include "plateauDeJeu.h"
-#include "Headers/nomdejoueurs.h"
+#include"Headers/play_cards.h"
+#include "Headers/hidden_cards.h"
+#include "Headers/cards_to_play.h"
+#include "Headers/player.h"
+#include "Headers/party.h"
 
-/// <- menuDeJeu.h
-void menu(int w,int h,int nbDePartie)
+
+//#include "menu.h"
+#include "aid.h"
+#include "scorescreen.h"
+
+#include "board.h"
+//#include "Headers/nomdejoueurs.h"
+
+
+///exit button
+volatile bool Fermer = false ;
+void Handler ()
 {
-    // Use current time as seed for random generator
-    srand(time(0));
+    Fermer = true;
+}
 
-/// counter to select from menu
+
+void menuDeJeu(void)
+{
+    Partie partie ;
+
+
+    ///exit si il tape sur x
+    LOCK_FUNCTION(Handler);
+
+    int w = 800;
+    int h = 600;
+    /// counter to select from menu
     int counter =0;
 
-    ///configuration de la police à partir du dossier source
-    //FONT *font1 =load_font("font1.pcx",NULL,NULL);
 
     /// double buffering screen
     BITMAP *menuScreen = create_bitmap(w,h);
+
     /// background image
-    BITMAP *menuImg = load_bitmap("assets/player_menu/menu_screen1.bmp",NULL);
+    BITMAP *BackGroundImage = malloc(sizeof(BITMAP));
 
-    //text to display for test
-    //textout_ex (backGroundImage,font1,"Hello World",round(w/2) -150 ,round(h/2),makecol(32,29,240),makecol(0,0,0)); //texte à afficher dans l'application
+    /// fermer le program
+    set_close_button_callback(Handler);
 
-
-    /// not exiting till esc is pressed
-
-    bool isDone= false;
-    while(!isDone)
+    /// ne pas quitter tant que esc n'est pas pressé
+    while(!Fermer)
     {
-
-
-        /// menu functionlity
-        if(key[KEY_ESC])
-        {
-            isDone=true;
-        }
-
-        /// logic de selection
-
+        /// Logic de selection
 
         if(key[KEY_DOWN])
             counter++;
 
         if(key[KEY_UP])
             counter--;
-        if(counter>3 )
-            counter=1;
+
+        if(counter>4 )
+            counter=0;
 
         if(counter<0)
-            counter=3;
+            counter=4;
 
-        /// anim les screen
         switch(counter)
         {
         case 1:
-            menuImg = load_bitmap("assets/player_menu/menu_screen2.bmp",NULL);
+            BackGroundImage = load_bitmap("assets/menu_screen/menu_screen2.bmp",NULL);
             if(key[KEY_ENTER])
             {
+                clear_bitmap(menuScreen);
                 /// -> partie.h
                 Partie newPartie ;
-                initPartie(&newPartie,2,nbDePartie);
+                initPartie(&newPartie,1,0);
 
-                clear_bitmap(menuScreen);
 
                 /// -> nomDeJoueurs.h
-                nomDeJoueurs(w,h,2,&newPartie);
+                //nomDeJoueurs(w,h,2,&newPartie);
+
+                strcpy(newPartie.joueurs[0].nom,"Lucky Player");
 
                 /// plateauDeJeu.h
                 plateauDeJeu(&newPartie);
                 rest(40);
                 set_gfx_mode(GFX_AUTODETECT_WINDOWED,w,h,0,0);
 
-                isDone=true;
-
+                /*
+                clear_bitmap(menuScreen);
+                menu(w,h,1); //prv :    menu(w,h,nb+1);    use nb+1 when  u finish the "work on this later"  Line 39
+                rest(50);
+                */
             }
 
 
             break;
-        case 2:
-            menuImg = load_bitmap("assets/player_menu/menu_screen3.bmp",NULL);
+        case 2:// partie saved
+            BackGroundImage = load_bitmap("assets/menu_screen/menu_screen3.bmp",NULL);
             if(key[KEY_ENTER])
             {
-
-                Partie newPartie ;
-                initPartie(&newPartie,3,nbDePartie);
+                partie = readPartieData();
 
                 clear_bitmap(menuScreen);
 
-                /// -> nomDeJoueurs.h
-                nomDeJoueurs(w,h,3,&newPartie);
+                plateauDeJeu(&partie);
 
-                /// plateauDeJeu.h
-                plateauDeJeu(&newPartie);
                 rest(40);
-                set_gfx_mode(GFX_AUTODETECT_WINDOWED,640,480,0,0);
-
-                isDone=true;
+                set_gfx_mode(GFX_AUTODETECT_WINDOWED,w,h,0,0);
 
             }
+
             break;
         case 3:
-            menuImg = load_bitmap("assets/player_menu/menu_screen4.bmp",NULL);
+            BackGroundImage = load_bitmap("assets/menu_screen/menu_screen4.bmp",NULL);
             if(key[KEY_ENTER])
             {
-                Partie newPartie ;
-                initPartie(&newPartie,4,nbDePartie);
-
                 clear_bitmap(menuScreen);
-
-                /// -> nomDeJoueurs.h
-                nomDeJoueurs(w,h,4,&newPartie);
-
-                /// plateauDeJeu.h
-                plateauDeJeu(&newPartie);
-                rest(40);
-                set_gfx_mode(GFX_AUTODETECT_WINDOWED,640,480,0,0);
-
-                isDone=true;
+                aideMenu(w,h);
+                rest(50);
             }
             break;
+
+        case 4:
+            BackGroundImage = load_bitmap("assets/menu_screen/menu_screen5.bmp",NULL);
+            if(key[KEY_ENTER])
+                Handler();
+            break;
         default:
-            menuImg = load_bitmap("assets/player_menu/menu_screen1.bmp",NULL);
+            BackGroundImage = load_bitmap("assets/menu_screen/menu_screen1.bmp",NULL);
             break;
 
         }
+
         ///set background image for menu
-        blit(menuImg,menuScreen,0,0,0,0,w,h);
+        blit(BackGroundImage,menuScreen,0,0,0,0,w,h);
 
         //texte à afficher dans l'application
-        textout_ex (menuScreen,font,"Appuyer sur ESC pour retourner",0,h-10,makecol(32,29,240),makecol(109,158,235));
+       // textout_ex (menuScreen,font,"Appuyer sur ESC pour quitter",0,h-10,makecol(32,29,240),makecol(109,158,235));
 
         /// double buffering
         blit(menuScreen,screen,0,0,0,0,w,h);
-
         ///show mouse
         show_mouse(screen);
         rest (50);
@@ -139,6 +144,7 @@ void menu(int w,int h,int nbDePartie)
 
 
     }
+
 
 
 }
